pyproject.toml

[tool.poetry]
name = "tempest_extractor"
version = "1.0.0"
description = "Extractor for Tempest weather stations for Cognite CDF"
authors = ["Greger Wedel <greger.wedel@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["arrow", "cognite", "dataclasses_json", "requests", "tempest_backfiller", "tempest_client", "websocket"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
cognite-sdk-core = "^2.56.1"
cognite-extractor-utils = "^2.2.0"
websocket-client = "^1.4.2"
dataclasses-json = "^0.5.7"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.7"
macholib = {version = "^1.14", platform = "darwin"}             # Used by pyinstaller pn Mac OS
pywin32-ctypes = {version = "^0.2.0", platform = "win32"}       # Used by pyinstaller on Windows
pefile = "^2019.4.18"                                           # Used by pyinstaller on Windows
mypy = "^0.971"
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"
pre-commit = "^2.20.0"

[tool.poetry.scripts]
tempest_extractor = "tempest_extractor.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
